requirements.txt
gunicorn
gevent
flask
Pillow
xmltodict